Let the hive UDF handle missing values same way as the model would
This PR changes the evaluate method to return actual predictions instead of a NULL if there is a NULL in the inputs.

According to the documentation, trees handle missing values in the input data by assigning them to the left branch of a split, i.e. they are a "valid" inputs for the tree classifier. The DRF and GBM expects these missing values to be 'Double.NaN's, hence the change. I'm not sure though if this is consistent among all the different model classes.